Add code and starting perldoc for warnings::fatal_enabled.
This is an analog for warnings::enabled, except it tests whether the
given category has been set fatal using "use warnings FATAL => foo".
This is mostly for symmetry.

Assumes that the fatal bit for a category will have an offset one higher
than the regular bit for the category, because otherwise much rewriting
of __chk would be required.